+++ src/scraper/types/output-target.class.ts
@@ -2,6 +2,7 @@ import { TargetType } from "./target-type.enum"
 
 export class OutputTarget {
   cssSelector: string
+  attribute?: string
   type: TargetType
   value: string | number | null
   name?: string