+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Make `ScrapeCommandDto.csr` actually work by supporting SSR-only mode using JSDom
+- Add `InputTarget.attribute: string` option to scrape the target's attribute instead of textContent
 
 ### Bugfixes